+++ src/aihwkit/extension/__init__.py
@@ -10,6 +10,8 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+# pylint: disable=import-error, no-name-in-module, invalid-name
+
 """AIHWKIT extension """
 
 from importlib.util import find_spec
@@ -18,3 +20,6 @@ EXTENSION_COMPILED = find_spec(".aihwkit_extension", package="aihwkit.extension"
 
 if EXTENSION_COMPILED:
     from .functions import FloatPrecisionCast
+    from aihwkit.extension.aihwkit_extension import ops as extension_ops
+else:
+    extension_ops = None